+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+
+## [X.Y.Z] - 2024-xx-xx
+
+### Features
+
+- CLI: allow multiple operations to be supplied at once for access control [#155](https://github.com/Cosmian/kms/pull/155).
+
 ## [4.11.2] - 2024-01-23
 
 ### Ci